Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 1088ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 240222/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -1,20 +1,20 @@
 #include <iostream>
 
 int main() {
-    int m, a, b;
+    long long m, a, b;
     
     scanf("%d",&m);
     scanf("%d %d",&a, &b);
 
-    int min = m;
-    int max = 0;
+    long long min = m;
+    long long max = 0;
 
     for (int i=a;i<=b;i++)
     {
-        int mid;
-        int s = 1;
-        int e = m;
-        int count = 0;
+        long long mid;
+        long long s = 1;
+        long long e = m;
+        long long count = 0;
 
         while (s<=e)
         {
